+++ app/src/main/res/layout/feeditemlist_header.xml
@@ -189,4 +189,47 @@
         tools:visibility="visible"
         tools:text="Updates disabled" />
 
+    <LinearLayout
+        android:id="@+id/descriptionContainer"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:padding="24dp"
+        android:background="?android:attr/colorBackground"
+        android:orientation="vertical"
+        android:visibility="gone">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="8dp"
+            android:textColor="?attr/colorAccent"
+            android:text="@string/state_deleted_not_subscribed" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="8dp"
+            android:text="@string/description_label"
+            style="@style/TextAppearance.Material3.TitleMedium" />
+
+        <TextView
+            android:id="@+id/descriptionLabel"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:ellipsize="end"
+            android:lineHeight="20dp"
+            android:maxLines="3"
+            android:layout_marginBottom="16dp"
+            android:background="?android:attr/selectableItemBackground"
+            style="@style/AntennaPod.TextView.ListItemBody"
+            tools:text="@string/design_time_lorem_ipsum" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/preview_episodes"
+            style="@style/TextAppearance.Material3.TitleMedium" />
+
+    </LinearLayout>
+
 </LinearLayout>
