+++ app/src/main/res/layout/onlinefeedview_activity.xml
@@ -20,6 +20,13 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
 
+        <ProgressBar
+            android:id="@+id/progressBar"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            style="?android:attr/progressBarStyle" />
+
     </androidx.cardview.widget.CardView>
 
 </LinearLayout>
